+++ CHANGELOG.md
@@ -8,6 +8,7 @@ List of changes for this repo, including `atomic-cli`,
 - Removed all HTML rendering from `atomic-server` (since we're using `atomic-data-browser`).
 - Changed how config paths are calculated and shared.
 - Remove the need for having the `./static` folder #118 when running `atomic-server`, moved to config dir.
+- Add `open config` to tray icon
 
 ## v0.23.0