+++ CHANGELOG.md
@@ -9,6 +9,7 @@ List of changes for this repo, including `atomic-cli`,
 - Changed how config paths are calculated and shared.
 - Remove the need for having the `./static` folder #118 when running `atomic-server`, moved to config dir.
 - Add `open config` to tray icon
+- Updated `atomic-cli` path, no longer requires quotes
 
 ## v0.23.0